refactor: move shaders location to own module
